Upgraded all deps from package.json.
Fixed typescript errors that new tsc revealed in our code as well as by
using newer angular.d.ts.
Tests were also failing now due to wrong order in modules & due to one
test expecting it's to execute in sequence. For now I disabled random
execution in Jasmine to keep that large test code separated in "it"'s.
If we want to change this back to random (maybe would be nicer) we have
to make one big "it" in foundsettype.test.js file.