SVY-12585 ValueList refresh mechanism on components not optimized
only in filling by foundset mode we will check the first (most parent)
related foundset to see if that is exactly the same instance
If that is the case we will not refill it.

So this only works for the most part for valuelist based on 1 relation
(not nested) or if it is nested it has to be filled based on foundset.
not by query. Based on query happens when the nested relations are based
on the same server and we dont display a calculation,