For SVY-12674 : ag grid grouping with valuelist
Refactored a bit to make foundset linked type wait for what it needs
before initializing itself - even when it comes from Rhino - so as to
not depend on the order in which properties get assigned / attached ...
even then. Custom object type code is no longer aware of foundset linked
type.