Fixed SVY-12602: Preformance in listviews [Autoflex]
Fixed/enhanced multiple things for webclient tableview with scrolling
enabled through application.putClientProperty and keep loaded records ==
false:
- data being initially sent twice in the same response and flickering
visually before scrolling to selection; first it was shown, then it was
removed/shown a second time then it scrolled to selection. Now it will
send data only once, keep the whole thing hidden until all is ready and
show/scroll/adjusting empty spaces top and bottom happen at once (the
flicker is gone).
- selecting any row when viewport started at idx > 0 (and was not close
to last record) would make that row jump to the top; fixed that, now if
it is visible it stays where it was before; if you click on partially
visible row on top or on bottom the scroll is minimally adjusted to show
that row + a few more pixels to be clear if there are more rows
preceding/following it; same goes for changing selection using up/down
arrows. In general if the new selected row is below visible area it will
be shown on bottom and if it is above visible area it will be shown on
top after auto-scroll.
- selecting any visible row when viewport start index > 0 would still
result in a flicker; it would send over from server to browser the whole
new viewport (centered on new selection). Now it sends nothing if
selection is already one of the rendered rows; so that avoids both the
flicker and unnecessary network traffic.