SVY-12724 View Foundset (readonly) support at runtime
made ViewFoundSet implement ISwingFoundset (so it is also a tablemode)

refactored FormControllers (and some others) to use only
IFoundsetInternal or ISwingFoundset. So that the ViewFoundSet can be
used in a Form (through solution model at this time)

To be able to use a ViewFoundSet in a Form you have to create one
databasemanager.getViewFoundSet() but also register
databaseManager.registerViewFoundset() before showing the form.
Also the form has a special loadRecords(ViewFoundset) that can be used
afterwards.