For SVY-12719: List FormComponent test and fix deleted/inserts of rows
First part of implementation that will help get on the client and in
component listeners a 'merged' list of update operations
(insert/delete/update) which will have their indexes match the indexes
in the foundset prop/foundset linked prop data viewport after all of
them are already applied. It should be much easier for the component to
use this sequence of merged events to granularly update it's display.
For more info please read javadoc of ViewportChangeKeeper class.