Implemented SVY-12719: List FormComponent test and fix deleted/inserts of rows
To keep the code of this and other components that are build on the
foundset simpler I refactored the way granular updates are sent from
server (merged and sorted). Now on the client you are guaranteed to get
a series of updates/inserts/deletes that, if you process sequencially in
the UI you are able to use the indexes from those in to get the data
from the already-changed viewport; before those indexes might not have
matched the ones from the final data viewport that is available at that
time.

The component itself is now able to handle insert/delete as well as get
updated properly when viewport startindex/size change making it
necessary to automatically change page or automatically load more
records.