SVY-12901 Update able ViewFoundSet support
implemented save([record]) on a foundset and viewrecords do keep a
change map and register themselfs by the viewfoundset as editting..

If there are editing records then all load() queries are postponed
except the one from scripting. That will discard all edits.

So the refresh from databroadcast but also the loadMore will not happen
if there are editted records.