SVY-13032 Create a PersistIndex class that hold the uuid,id,name indexes of the FS
idToPersist has now an indirection to Persist.class->id->persist
this way for Form,Relations,ValueList and Media we can way quicker get
to the actual list, instead of walking through a huge list of all global
objects.
ScriptVariable and ScriptMethod are still walking of the all objects
list because they are meant to return only the globals.