For SVY-13186: AGGrid global valuelist with return value text isn't working
Valuelist type should not mark itself (or it's wrapping foundset linked
prop) as changed when it's items change but it is not sending items
anyway to client. This combined with the fact that dataprovider property
triggers changes due to auto-resolve of real values during it's actual
JSON rendering could lead to valuelist fs linked property remaining in a
wrong state (thinks it has changes and monitor knows about it but
monitor actually cleared it's changes) not sending new changes next time
it needs to.