SVY-13440 Error when createing in memory dataset in onLoad of form
make sure that the default PersistIndex doesn't listen to any persist
change events. Because it should not update changes a client makes
through the solution model.
But in solution model and also in the developer persist index we do need
this to update the cache.
For client in de developer, we then do need to rebuild the cache when
the persist is changed.