SVY-13440 Error when createing in memory dataset in onLoad of form
give the newParent right away to the clone and add the clone before
filling it fully first to the new parent. so get
getRootObject().getChangeHandler() really works on the right object
where the clone is really added to.
else it would fire all the changes to the original solutions change
handler (where nothing is really changed for)