SVY-13440 Error when createing in memory dataset in onLoad of form
clone should not add the child if the newparent is the same as the old
parent. Then the clone should be fully completely yet (and the uuid
should be reset)
If it is a new parent then it should be added,so that the fire will go
to the right stuff, the resetUUID will update he uuid of the child and
in the uuid map of the parent.