Fix attempt for SVY-13936: Wrong value saved when you select a value from a Combobox that is still loading
When it tries to highlight the options ctrl.open is already false and an
ng-if has actually already hidden all the dom elements for the items. I
could reproduce that twice but very hard - and can't reproduce it
consistently to figure out why that happens. But as that is a $timeout
it makes sense to check the state before doing anything in it.