SVY-14051 datagrid updating problem
made sure we register the global variables in a FoundsetDAL so that when
we get an event for that specific global var, we will flush the whole
viewport of the full foundset so that the related data on all the rows
are updated.
It would be nicer if we could really only update the columns viewport
and only fallback to a full viewport change when needed.