Related to SVY-14172
Not the actual problem in the case but something I noticed when testing:
test solution had initially a body part that had a list form component
the body part had height 372 let's say and list form component was css
pos 0, 0, 0, 0 with height set to 360 let's say that resulted in markup
of form minHeight 372 and markup of list form component minHeight 360;
then I removed the 360 from list form component minHeight for that went
away; then I dragged body part to height 150 let's say and markup had
minHeight 150. Everything is as expected so far but then I wanted to go
back to initial sizes to check something and I did that but form
minHeight stayed 150, although visually it was 372; body part height was
372. I checked the .frm and when I dragged initially it set one extra
prop. at form level size: 639, 150 and then when I drag it back
that size remains the same just body part gets increased to 372.

Basically form.getSize() does compute the height correctly but the heigh
part of the size that is stored in a form persist must always be ignored
as it is not relevant.