SVY-14219 introduce forms.myform.containers.name.addStyleClass/removeStyleClass()
runtime support for adding/removing and testing for style classes
If form is responsive then a ContainersScope will be added to the
FormScope.
The ContainersScope returns a ContainerScope when the name is found that
has a container with a name.
That one dynamically provides the 3 methods an redirects the calls to
the FormUI that pushes it into the properties of the form itself
Then the client will pick up those properties and adjust the style class
of the container of that form accordingly