Fixed SVY-14993 : Audit login not working on Servoy Cloud
Problem: if it was a .dbi based wat export it would not export the DB
security info. If it was a .dbi export and DB was not available for
.servoy export the same would happen. But it did work if you did a
non-dbi based export in both cases or if for .servoy it was a .dbi
export but with DB available.

Now it should work in all cases (if it's a .dbi based export it no
longer looks based on DB but on .dbi and .sec files for those .dbi
files, if it's non-dbi export it looks at the DB).

Conflicts:
	servoy_shared/src/com/servoy/j2db/util/xmlxport/IXMLExporter.java