Fix for root cause of SVY-14914: Angular Digest (25) loop when having a labelFor on a table view (toggle visible)
Actual problem was that we were no removing portal.js listeners
registered to childElements and to relatedFoundset when $scope got
destroyed - which remain in form's model area that is reused when form
is hidden/re-shown. But hide of form would destroy scopes and elements
of portal and then we ended up on next show of the form which builds new
scopes and elements with old listeners digesting old destroyed scopes
which lead to very weird errors + memory leak.