SVY-15069 Very slow application. After Tomcat restart the issue is solved
implemented it that the message queue will be flushed at the client
side, that the server Proxy will just call 1 processQueue method with
all the differnet messages that are then processed client side.
The side effect is that we will not constantly check if there is more in
the queue, so no other client will get around, this process message will
be called once and then it gives the pool back to another client even if
it did get new messages in that time.