SVY-15189 duplicate uuid generation when cloning through solution model
removed the interace method without a parent argument
so we are forced to give a parent (still can be null)
used it in as many places as possible to directly give the parent
This way the parent structure is alread the cloned one before
getChangeHandler() is called on it. So it doesn't hit the
DeveloperPersistIndex but just the solution model one.