For SVY-15236: Upgrade Production Server to 2020-06 Failing
I think this case is finally fixed now - by this commit.
It was starting to load Solution "A" which had a form that used a Style
"A". Because of the same name and because it didn't check based on root
object type as well then the thread would wait for "A" to be loaded (the
style) which actually waited for itself to load solution "A" => hang.

Conflicts:
	servoy_shared/src/com/servoy/j2db/persistence/RootObjectCache.java