SVY-15779 valuelist are being cleared from memory?
removed the whole method setValueListItems method in ngclient to just do
what the base does, which already had "false" for the use softCache
changes that we now have is that NGClient uses now softCache, it does a
loop of form.refreshView() which is not needed but thats an empty
method. the guessed type is only set when it is not OTHER
for other clients setRuntimeChanged is also called but not used anyway
later on.