SVY-15738 introduce "clientfunction" property type to avoid doing "eval(model.xxx)"
Introduced the ClientFunctionPropertyType that for NG2 just generates a
"UUID" value which maps then on a generated js file that has all the
client functiosn as a map of that uuid->function

The NGClient does have a map of client side functions that and it will
call a client service to reload the stuff when a new one is added to
that map.

The servlet mapped on /clientfunctions.js will then serve out these
functions.