Only dataproviders should do a Apply/SvyPush.
if it is a dataprovider type then the datachange callback should know
this so it can send an "svyPush/apply" to the server instead of just a
"dataPush". Add a boolean argument to the callbacks.datachange

also ignore for now all depricated components, very likely we are not
going to port those.