SVY-15944 NG2 StyleClass is not applied to extended forms
made sure that FormWrapper uses getFlattendPropertiesMap() instead of
just its properties map.This seems to be never a problem in NG1 but NG2
really uses all those properties as the actual model.

FlattendForm has its own getFlattendPropertiesMap() so that doesn't
constantly have to go over the whole tree filling a map that can be done
at once in the fill method.