SVY-15989 Dynamically clone DB Servers or pair 1 connection with 1 client
Many interface changes for adding a client connection definition.
datasources.db.server has now a defineClientConnection() method.
So that the  Server.getDataSource() can return now a client configured
connection
(username,password or connection properties can be changed)
The server has gained a property "clientOnlyConnections" where we try to
postpone connecting to the server (also for getting meta data, like the
tables). Only do that when a client defintion is created.