SVY-16249 Make PerformanceData fully concurrent
removed all synchronized stuff as possible (only SortedList are synced
but much smaller sync block, just the list itself)
Made all the maps concurrent.
Made all fields final
moved the Aggregate to its own thing (so Data timing and aggregating are
not 1 thing)
Removed the subactions stuff, now all actions are based on stack. (and
are auto "sub"). All PerformanceData now report itself to the main
aggregate.
made sure that in a few places the clientid is taken and stored in a
field to use with start and end (because the client could shutdown and
not have a clientid anymore)