SVY-16249 Make PerformanceData fully concurrent
moved the start/stop performance timing to a function wrapper
This way all functions are profiled (that has a bit more overhead)
but the calculation of the name and the getting of the peformance data
object is only done once per function. (not per call)
In production where this profiling can be turned off then it shouldn't
have any effect. (no wrapper is created, this does result in that a
client must be restarted when it is client performance is enabled)