Fixed SVY-16385: foundset null pointer
A change in a scope variable made the foundset used by a component based
on a related selector relation1.relation2 become null while also
changing the record that used to be in that foundset property's DAL (due
to calculations). And that record that was now obsolete was triggering a
change in the DAL that now had a null foundset.

Fix: obsolete record should be removed from DAL when foundset is updated.