SVY-16425 Perfomance problems with foundset.loadRecords()
make sure the viewport is not tried to be calculated from a wrong
selection.
the state in the foundst is not fully corrected when fires for deletes
are done, the selection is then set after that.
a small fix to better detect this is a loadRecords() call (deleted
exactly from chunk size and the foundset size is chunk size)
still this is not perfect because potentially a normal delete could
match that exactly