For SVY-16278: Serverside logic to resolve the form/elementName for a JSEvent-typed property wrong
Fixed a regression caused by previous commits on this case.
A smartdocumenteditor component was using a 'function' typed value from
.spec to call extecuteInlineScript; the targeted method was a scope
method - so formname in the execute call was null (so the context in
fromJSON of JSEventPropertyType was root form), but formName in the
JSEvent was correct (the most deeply nested form that triggered the
event).

So fix was to use what JSEvent creation does detect client side for
JSEvent again.