Fix for SVY-16509: Exception when trying to edit a cell in nggrid
I suppose it is possible that a fromClientToServer comes in between the
beginning of fromServerToClient (that does removeWatches and sets that
to null) and end of fromServerToClient (that does addWatches and sets it
back to {}). That means that a child smart property notifies that it has
changes right away after updates from server... In that case the check
still means that changes are not in dumb property so the null check
should be ok (it goes to else which is smart props.).