SVY-16383 changeListener for FS in WebComponent called with viewportRowsCompletelyChanged when there is no change
If the PksAndRecordsHolder gets a new dataset, then it will compare the
pks with the previous set if it had one (only for pk sets that are 1
column).
It generates a list of changes based if the pk did change (change,
first, last) or when there where more rows (insert, first,last) or less
rows (delete, first, last)
Then foundset.fireDifferences() will use that instead of just looking at
the 2 sizes. This way it will fire a lot less if the foundset was not
really changed, Buf it 1 pk changed somewhere (inserted) then from that
row on it will be changed fully

added property "servoy.foundset.optimizedChangeFires" (default true)