SVY-16572 Slow developer with some forms
cache it in the Form and add a method getSuperPersist(id) that will
return or the forms component itself with that id or the component that
has that also as an extend id.
make sure that a form knows when something changes so it can flush its
cache. (at runtime this never happens for the none solution model
solutions)