Fix for SVY-16948: ds.createDataSource throws error in Servoy 2022.03RC1
1. As fixedColumnTypes == null is a valid scenario, make sure it does
not cause a NPE (it was problematic where this exception happened and in
another place related to rowsToStringserialize).
2. In one place where it checked inserted row count it used fixedDataSet
(that might be a 0 row dataset meant only to generate the correct
columns for the tables if given columns don't match) instead of dataSet.
3. When it added only partial column content to the table, the column
types given need to match (by index) the data from the dataset. So
columnTypes instead of fixedColumnTypes for dataSet.