For SVY-17013: Layout problem with anchored form component inside responsive main form
A comment on the case made me test some more scenarios. So basically we
only use min-width and min-height instead of width and height + 2 more
situations:
1. where anchored in responsive when one wants to set width should also
add float: left (it assumes you want to add multiple items in one 12grid
column for example)
2. when anchored in anchored, it will add height: 100% as well in order
for bottom anchored components to work as expected as well.