Fixed SVY-17129: Issue in WebComponent: receiving model property changes due to RhinoConversion?
Because we want to generate a NativeArray that is not dense, we give a
big initial length, and then set length on it to what it is supposed to
be. That does generate delete ops. as well during initialization that
should not trigger change by ref on the sablo prop. So ignore (do not
register changes for) deletes as well during init, not just puts.