For SVY-17166: Make hiding forms clearer - even with nested tabs
- StopIfEditing is now called similar to onBeforeHide - so it goes over
all children, checks stopIfEditing and onBeforeHide to see if hiding is
allowed before actually doing the hide/onHide. Before it would only care
about what stopIfEditing returns in the first form that is hidden, not
in visible child UI forms - and those could remain thinking that they
blocked the hide and are still visible even though the system did hide
all those forms.
- a small change in DAL - where it is called from WebFormUI.notifyVi...
which is not a pre-hide step, it should give false as arg.
- removed the "return true" template when creating a onHide method -
return value is now deprecated.