SVY-17304 setting log level and tracing doesn't work in admin page (from developer)
Configurator.setAllLevels can only be called by a name not class
and if you do that in another callee classloader that classloader is
used (like the server plugin classloader of admin page)
then the Debug class logger itself doesn't see this change.
You need to call that on the same classloader as where the logger is
created.