Fix for SVY-17441: Cannot login in sample solution
Form from WebFormComponent.visibleForms, that was already hidden and
destroyed - probably by
FlattenedSolution.clearLoginSolution(IActiveSolutionHandler,
IServiceProvider) - which does not care about the UI nesting order of
visible forms when hiding and destroying them, nor does it update the
WebFormComponent.visibleForms member of parent visible forms if it
hides/destroys a child before doing it on parent - was already
hidden/destroyed. And that caused the NPE. Just allow "hide" in that
case without doing anything else.