Fixed SVY-17429: Session without websocket are not closed and hangs on server
"Sablo Session closer" thread remained forever waiting for an
NGClient.invokeAndWait() to execute - that was supposed to shut down a
client who's session has expired after websocket was lost for >60 sec
(tab browser close).

Problem was that it ended up posting that shutdown event on an already
destroyed event dispatcher that would no longer process the event. Even
the even processing thread was gone.

From then on, no more client shutdowns could happen on that server via
websocket/session timeout. So basically the server needed a restart.

See case comment for a detailed description of how it could be
reproduced.