SVY-17440 Default "latest" theme does not give an NG2 version
made the ng2 theme properties insync with the ng1 (so it now has those
doc behind with the categories)
this could also mean that there is a slight change in properties.
This needs to be resynced back to the theme roller
(created a follow up case)

removed from the versions the "_ng2" one now all the versions are just
versions plain versions (year.q.minor) and it will be just postfixed in
the file. or postfixed when we want the ng2 variant.

getResource() can now handle "latest_ng2"that will return the latest
version that there is postfixed with "_ng2"