For for SVY-17514: Regression: cannot set methodArguments on PopupMenu anymore
Keep the initial native object as well in the prototype chain. It is a
place to store any internal server-side state properties that are not
declared in the .spec file. For example window_server.js's
"menuArgumentsInternal" global variable could be completely removed now,
and have that stored in a menu item's "this.menuArgumentsInternal".