SVY-17612 Images at random missing in self made component
Dynamic Media resources are now stored in a client instance
They are only cleaned when solutionClose() happens.
So they now live as long as the client lives
They are still saved to disk if they are bigger then 5MB (through the
MediaResourceServlet)
Now they really can only be served by a real client only, so to get the
media you need the url and the http session.