SVY-17727 variants rewrite to use a json file that just is a set of classes
moved the VariantsHandler to the FlattenedSolution so the media is
always get from there.
Introduced a VariantPropertyType that converts a named variant to a
array of classes.

a fix in the childrenjsongenerated to give the context when writing the
json data (so the toTemplate call gets the right client/fs)