SVY-17656 deadlock can happen when a databroadcast happens at the same time as a saveData() on the same foundset/records
first go over the all the rows to test for a auto enter field or if it
is db managed. Then go over the rows to get the final values
This is because an auto enter column can be used in a stored calculation
that is then modifying a stored calculation.