For SVY-14140
Dataprovider type on server knows to expect a date conversion for date
DPs even if the value sent to client previously was just null (with no
dynamic type, so client didn't know type). So in case of DPs, the
applyDP should use the date converter directly for Date instances, not
the ObjectConverter.