SVY-17621 Issue with dataprovider type lookup for format properties in Webcomponent spec
use FoundsetTypeSabloValue.getTableBasedOfFoundsetPropertyFromFoundsetIdentifier(foundsetId,
application, form);
as much as possible because that will be able to return the right table
for almost all scenario's: form foundsets, relation, datasource string
(db:/server/table) but also ViewFoundset datasource strings.

if the table is not found then try to fallback and use the parent
context of the webContext, try to lookup there the foundset property
type.