Fixes for SVY-17654: Upon foundset load grid is showing wrong data
In one situation, viewport bounds were not corrected after a delete
event - if the foundset shrinked into the viewport's rows. Also
refactored a bit the insert code because the change notifications and
marking took a long time to understand there - code in 2 classes
depended on each other in a hidden way (that send bounds only + didn't
seem to call the notify method but then again it didn't need to because
of the other code...).