For SVY-18016: Global Method Valuelist called with wrong row
Some code that was trying to call queueCellChange later (because, before
the refactor for case SVY-14140, it used to call toJSON directly on that
cell and that did FoundsetDAL.setRecordQuietly which would revert any
corrections that were done server-side to the DP value received from
client - because those changes were not yet pushed to the record), after
a DP apply would both set the new value and push it to the record was
now obsolete and is removed, because now queueCellChange only keeps
track of changed indexes/changes and no longer does
FoundsetDAL.setRecordQuietly nor does it do toJSON right away, but later
when the changes are really written for the viewport.

It will also not reset the FoundsetDAL to selected record fast after
applying updates from client anymore - if viewport changes are to be
written; because that reset to selected will happen anyway when viewport
changes are sent to client. And in case of a valuelist.filter on a
non-selected record it caused problems because the filter was lost due
to the record changed that happened before toJSON was called on the
valuelist property.