Fixed SVY-18133: NG1 Error with Form Container & Advanced Render
The 'data' property of foundsetlist in custom renderers package was
marked in .spec as scope: private and it had default pushToServer which
is reject.
That means that it was never sent from client to server nor from server
to client; it was just assigned on client. So the value which was a
custom array was never initialized to have an internal state.

But now (after SVY-14140) that the client knows the property types it
did call updateAngularScope (NG1) on it - which had to be able to deal
with an uninitialized internal state.