SVY-18323 In NG(1) application.getActiveWindow() doesn't work properly
NGRuntimeWindows will not maintain also a childrens list of all the
dialog they show (they are a parent off)
If they get closed all the children (and the children of he children)
are closed first, if onHide() fails it will stop the closing of the
parent.
Also destroy() will first call hide() and if that fails destroy will be
blocked.
When scripting is run outside of a ui event, like a scheduler or an
async request, the getActiveWindow() will now return the toplevel modal
dialog if the current window is a parent of that.