SVY-18286 Ability to type certain properties in a property of type 'map' (in Web Components)
introduced a IDesignMapValueConverter that is implemented by
ClientFunctionPropertyType and the Map (Json) type itself

This has a method to create a json object that will be converted on the
server and client to the right thing.
It now has support for nested maps/json structures including json
objects and arrays.

updated docs for:
https://docs.servoy.com/reference/servoycore/dev-api/property_types#clientfunction
https://docs.servoy.com/reference/servoycore/dev-api/property_types#map
https://docs.servoy.com/reference/servoycore/dev-api/property_types#json