SVY-18384 labelFor does not work with form components
a bit diffent implementation to get to the right component.

see inline comments

This still doesn't fully resolve it nicely when having many form
component components with the same form that all also have label for and
so on
And should a form component point to a normal component on a form and
the other way around?
or should a form component label of an FC1 be pointing to a FC2
component?