For SVY-18716: Combination of modal dialogs and plugins.svyBlockUI doesn't work properly
I tried in many ways to reproduce SVY-11302 in NG1 on current master but
I couldn't. So I don't know what was really the problem there or if that
problem still affects Titanium (for the cases with form visibility and
form load I was able to reproduce the old cases and see that Titanium is
not affected by those). That is why I will leave the formUnload just as
it was before; even though I wanted to add that to the if NG1 so that
Titanium no longer executes these events that can trigger handlers in an
unexpected order due to sync calls to client.