For SVY-18721: element attributes property cannot be modified in place
The FormElement value of a ServoyAttributesPropertyType property ended
up directly in the client - and could be manipulated by solution code;
however this is shared via the form element cache. So the value of
'attributes' on a component that was changed at runtime by one client
ended up being visible/used in another client.

This does not happen with other types or primitives because they either
can't have a design value (if it's not a primitive) like 'object' or
they have a conversion from form element to sablo that changes instance.
Primitives do not count because they always change by ref.