SVY-18833 make "function" type a callable function so that it can be used at once and as callback functions directly in 3rd party components
use the FunctionRefType so that is also just callabale
now stuff like this works:

	elements.inputgroup_1.map = {test: function(ev, name, text, index) {
		application.output(ev);
		application.output(name + " ::: " + test + " ::: " + index);
	}}

for map, json and object types.

The function sends over its a hash that is kept as a weak reference and
that is now callable with the same executeInlineScript call but that is
also wrapped in a function.
So there are now 2 modes, 1 is form/script which must point to a top
level servoy function (and can also be selected in the designer if it is
a "function" property.
The second one is as above can be any kind of function also inner/inline
functions