SVY-19120 TiNG clienside functions in ChartJS
keep wrapping all the maps as json objects (nativeobject types)

Not sure why this was not happening before but if you have a nested
structure:
	var plgins = {
		tooltip :  {
			callbacks: {
				label: application.generateBrowserFunction("function
blaat(){console.log('test');}")
			}
		}
	}

and that is assigned to a json property then it would end up in a stack
overflow because it tries to "json" the BrowserFunction object.
which results in a stackoverflow.

This is because the fixJSONObjectValueTypesAndI18N doesn't expect that
the nested json objects are still NativeObjects instead of JSONObjects
and then the serialzie to the client will go wrong because of the
BrowserFunction