SVY-19496 Screen with 3000+ fields on it (Custom advanced search) will never load in TiNG
only write out the position when this is a anchored form where
components don't have css position (then convert that)

for responsive forms the (css) position doesn't have to be send at all.
for css position forms it it send through the model, the client side
will pick it from there.