For SVY-19635: Foundsetpush issue
The new foundsetInitialPreferredViewportSize property type should also
set centerOnSelectedRecord to false - in order to send paging like
viewport initially or when auto-send is done. The old option to center
on selection was only available in client-to-server
setPreferredViewportSize call... I made it available both in the .spec
file and when setting a new foundset prop. from rhino.