For SVY-19635: Foundsetpush issue
Fixed a failing test (that was too rigid in how it would look for
service calls that are sent to client); made the whole thing more
flexible (not it can look easier for a specific call in all the messages
so far, look for it's first, second etc. occurrence, check or not the
args).

This way I was able to easily add more checks to see that multiple
different calls are sent to the client in the expected order.