For SVY-19841
If a main form that has a related tab showing is hidden, and the form
from that tab is shown in an unrelated tab in the next main form, then
the "related" foundset show remain in that form - although it is
detached from the UI related listeners; so it will not update anymore
because the second main form does not say it's related data.

An e2e test was failing due to this as now the related tab's foundset
was set to null when parent UI form was hidden.