SVY-20085 FunctionDefiniton should be relaxed a bit so you can give arrow functions to it
now if a form/method names can't be found just remember the function and
call that one through the scriptengine
It is called tthroug invokeNow/Later to be sure that it is executed on
the event thread.

This is dangerous for form function that give arrow functions to
something and that can be done way later when the form is destroyed so
people need to be carefull with that.