SVY-20189 Invicti Security Scanner - Possible B.R.E.A.C.H. Attack Detected
added a property "servoy.ngclient.output.querystring.for.url.rewrite"
that is default false so it doesn't output the query string
For people that used rewrite urls in tomcat or in front of our filter
that maps a nice url to an url with arguments (but that is now shown in
the browser) thos need to set this property to true.

in the future we can maybe enhance this a bit by passing the query args
through the session to the websocket, but this has more problems, for
example when having multiply tabs that load at the same time.