SVY-19820 Handle bigints (java/javascript) to be able to send to the client (json doesnt directly support this)
added a utility function in JSUtils createLargeNumber where you can give
a string, this way you can create large numbers that are represented as
a BigDecimal in java. This is added because you can't create such
numbers directly in javascript as literals because they will be boxed
into a double.